Extract visitors to stand-alone classes
